+++ collect_app/src/test/java/org/odk/collect/android/utilities/QuestionFontSizeUtilsTest.java
@@ -9,6 +9,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.TestSettingsProvider;
+import org.odk.collect.android.widgets.utilities.QuestionFontSizeUtils;
 
 @RunWith(AndroidJUnit4.class)
 public class QuestionFontSizeUtilsTest {
