+++ collect_app/src/test/java/org/odk/collect/android/widgets/ExVideoWidgetTest.java
@@ -7,7 +7,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.odk.collect.android.utilities.QuestionFontSizeUtils.DEFAULT_FONT_SIZE;
+import static org.odk.collect.android.widgets.utilities.QuestionFontSizeUtils.DEFAULT_FONT_SIZE;
 import static org.odk.collect.settings.keys.ProjectKeys.KEY_FONT_SIZE;
 
 import android.view.View;
